requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
setuptools==69.2.0
sqlparse==0.4.4
tzdata==2024.1
whitenoise==6.6.0
wheel==0.42.0
django-imagekit==5.0.0
django-cleanup==8.1.0
